+++ RestSharp/Serialization/Xml/DotNetXmlDeserializer.cs
@@ -1,37 +1,31 @@
-﻿#region License
-
-//   Copyright 2010 John Sheehan
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License. 
-
-#endregion
-
-using System.IO;
+﻿using System.IO;
 using System.Text;
+using RestSharp.Serialization.Xml;
 
-namespace RestSharp.Serialization.Xml
+namespace RestSharp.Deserializers
 {
     /// <summary>
     /// Wrapper for System.Xml.Serialization.XmlSerializer.
     /// </summary>
     public class DotNetXmlDeserializer : IXmlDeserializer
     {
-        public string DateFormat { get; set; }
-
-        public string Namespace { get; set; }
-
+        /// <summary>
+        ///     Name of the root element to use when serializing
+        /// </summary>
         public string RootElement { get; set; }
 
+        /// <summary>
+        ///     XML namespace to use when serializing
+        /// </summary>
+        public string Namespace { get; set; }
+
+        public string DateFormat { get; set; }
+
+        /// <summary>
+        ///     Encoding for serialized content
+        /// </summary>
+        public Encoding Encoding { get; set; } = Encoding.UTF8;
+        
         public T Deserialize<T>(IRestResponse response)
         {
             if (string.IsNullOrEmpty(response.Content))
@@ -39,7 +33,7 @@ namespace RestSharp.Serialization.Xml
                 return default(T);
             }
 
-            using (var stream = new MemoryStream(Encoding.UTF8.GetBytes(response.Content)))
+            using (var stream = new MemoryStream(Encoding.GetBytes(response.Content)))
             {
                 var serializer = new System.Xml.Serialization.XmlSerializer(typeof(T));